requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
jupyter>=1.0
pyyaml>=6.0
dash>=2.14
flask>=2.2
